requirements.txt
streamlit
openai
pandas
numpy
scikit-learn
pillow
opencv-python-headless
pypdf
pytesseract
sentence-transformers
faster-whisper
faiss-cpu
qdrant-client
matplotlib
transformers
pymupdf
pyarrow
python-dotenv
